b-ai-tutor/server/requirements.txt
Flask>=2.2
flask-cors>=4.0
redis>=4.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.1
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
pdfplumber>=0.9.0
werkzeug>=2.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.0.0
